+++ vpin-studio-rest-client/src/main/java/de/mephisto/vpin/restclient/dmd/DMDDeviceIniConfiguration.java
@@ -1,13 +1,18 @@
 package de.mephisto.vpin.restclient.dmd;
 
 public class DMDDeviceIniConfiguration {
+
+  public static final String WEBSOCKET_DMD_PATH = "/dmd";
+
   private int emulatorId;
   private boolean networkStreamEnabled = false;
   private String webSocketUrl;
-
+  private boolean webSocketUrlInvalid = false;
+  
   private boolean useRegistry = false;
   private boolean stayOnTop = false;
-  private boolean enabled = false;
+  private boolean ignoreAspectRatio = false;
+  private boolean enabled = true;
 
   public boolean isStayOnTop() {
     return stayOnTop;
@@ -17,6 +22,14 @@ public class DMDDeviceIniConfiguration {
     this.stayOnTop = stayOnTop;
   }
 
+  public boolean isIgnoreAspectRatio() {
+    return ignoreAspectRatio;
+  }
+
+  public void setIgnoreAspectRatio(boolean ignoreAR) {
+    this.ignoreAspectRatio = ignoreAR;
+  }
+  
   public boolean isEnabled() {
     return enabled;
   }
@@ -55,5 +68,13 @@ public class DMDDeviceIniConfiguration {
 
   public void setWebSocketUrl(String webSocketUrl) {
     this.webSocketUrl = webSocketUrl;
+  }
+
+  public boolean isWebSocketUrlInvalid() {
+    return webSocketUrlInvalid;
+  }
+
+  public void setWebSocketUrlInvalid(boolean webSocketUrlInvalid) {
+    this.webSocketUrlInvalid = webSocketUrlInvalid;
   }
 }