+++ vpin-studio-server/src/main/java/de/mephisto/vpin/server/dmdscore/DMDScoreProcessor.java
@@ -4,7 +4,7 @@ public interface DMDScoreProcessor {
   
   void onFrameStart(String gameName);
 
-  void onFrameReceived(Frame frame, int[] palette, int width, int height);
+  void onFrameReceived(Frame frame);
 
   default void onFrameStop(String gameName) {}